+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.2.0 (DEV)
 
+- **BREAKING**: Remove generic type parameter 3 in `withPersistent()` call.
+- **BREAKING**: Remove generic type parameter 3 in `withPersistentMap()` call.
+- Fix: Improve type inference in `withPersistentMap()` call in some cases.
 - Upd: Interface `WithPersistentOptions` is now generic with 3 optional type
   parameters.
 - Add: Options `serialize` for serialization before writing to driver and